Updated folder structure
- Move _property_serialization, _utils and _types into view
- Move a couple of utils up to plugin level for ProviderSet
- Move ProviderSet and DerivedVectorAccessors from _types to _utils
