+++ webviz_subsurface/plugins/_prod_misfit/utils/make_dataframes.py
@@ -4,8 +4,9 @@ from typing import Dict, List
 
 import pandas as pd
 
-from ..._simulation_time_series.types.provider_set import ProviderSet
-
+from ..._simulation_time_series.views.subplot_view._utils.provider_set import (
+    ProviderSet,
+)
 
 # -------------------
 def get_df_smry(
