+++ webviz_subsurface/plugins/_property_statistics/controllers/property_response_controller.py
@@ -11,9 +11,10 @@ from webviz_subsurface._utils.dataframe_utils import (
     correlate_response_with_dataframe,
     merge_dataframes_on_realization,
 )
-from webviz_subsurface.plugins._simulation_time_series.utils.datetime_utils import (
+from webviz_subsurface.plugins._simulation_time_series.views.subplot_view._utils.datetime_utils import (
     from_str,
 )
+
 
 from ..models import (
     PropertyStatisticsModel,
