+++ webviz_subsurface/plugins/_simulation_time_series/_plugin_ids.py
@@ -4,8 +4,3 @@ class PluginIds:
     class SimulationTimeSeries:
         GROUP_NAME = "simulation-time-series"
         VIEW_NAME = "simulation-time-series-view"
-
-    # pylint: disable=too-few-public-methods
-    class TourStepIds:
-        DELTA_ENSEMBLE = "delta-ensemble"
-        OPTIONS = "options"
