+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_property_serialization/graph_figure_builder_base.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Optional
 import pandas as pd
 import plotly.graph_objects as go
 
-from ..types import FanchartOptions, StatisticsOptions
+from .._types import FanchartOptions, StatisticsOptions
 
 
 class GraphFigureBuilderBase(abc.ABC):
