+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_settings/_filter_realization.py
@@ -8,7 +8,7 @@ from webviz_config.webviz_plugin_subclasses import SettingsGroupABC
 
 from webviz_subsurface._utils.formatting import printable_int_list
 
-from ....types import StatisticsFromOptions
+from .._types import StatisticsFromOptions
 
 
 class FilterRealizationSettings(SettingsGroupABC):
