+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_settings/_time_series.py
