+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_types/types.py
