+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_utils/delta_ensemble_utils.py
@@ -2,7 +2,8 @@ from typing import Dict, List, Tuple
 
 from webviz_subsurface._providers import EnsembleSummaryProvider
 
-from ..types import DeltaEnsemble, ProviderSet
+from .._types import DeltaEnsemble
+from .provider_set import ProviderSet
 
 
 def create_delta_ensemble_name(delta_ensemble: DeltaEnsemble) -> str:
