+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_utils/derived_vectors_accessor/__init__.py
@@ -3,13 +3,3 @@ from .derived_delta_ensemble_vectors_accessor_impl import (
 )
 from .derived_ensemble_vectors_accessor_impl import DerivedEnsembleVectorsAccessorImpl
 from .derived_vectors_accessor import DerivedVectorsAccessor
-from .provider_set import ProviderSet
-from .types import (
-    DeltaEnsemble,
-    FanchartOptions,
-    StatisticsFromOptions,
-    StatisticsOptions,
-    SubplotGroupByOptions,
-    TraceOptions,
-    VisualizationOptions,
-)
