+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_utils/derived_vectors_accessor/derived_delta_ensemble_vectors_accessor_impl.py
@@ -11,8 +11,8 @@ from webviz_subsurface._utils.vector_calculator import (
     get_selected_expressions,
 )
 
-from ..utils import dataframe_utils
-from ..utils.from_timeseries_cumulatives import (
+from .. import dataframe_utils
+from ..from_timeseries_cumulatives import (
     calculate_from_resampled_cumulative_vectors_df,
     get_cumulative_vector_name,
     is_per_interval_or_per_day_vector,
