+++ webviz_subsurface/plugins/_simulation_time_series/views/subplot_view/_utils/from_timeseries_cumulatives.py
